move parallel dims to distributed folder
